pyproject.toml
[project]
name = "geosetter_lite"
version = "0.1.0"
description = "Geotagging application"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "PySide6>=6.6.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "PyYAML>=6.0"
]
